Some optimization in CleanupsBeforeDelete (#1536)
Credit MangosR2

+++ src/game/Unit.cpp
@@ -11085,11 +11085,13 @@ void Unit::RemoveFromWorld()
 
 void Unit::CleanupsBeforeDelete()
 {
+    if (!IsInWorld())
+    return;
+
     // This needs to be before RemoveFromWorld to make GetCaster() return a valid pointer on aura removal
     InterruptNonMeleeSpells(true);
 
-    if (IsInWorld())
-        RemoveFromWorld();
+    RemoveFromWorld();
 
     ASSERT(GetGUID());
 
@@ -11101,8 +11103,12 @@ void Unit::CleanupsBeforeDelete()
     m_Events.KillAllEvents(false);                      // non-delatable (currently casted spells) will not deleted now but it will deleted at call in Map::RemoveAllObjectsInRemoveList
     CombatStop();
     ClearComboPointHolders();
-    DeleteThreatList();
-    getHostileRefManager().deleteReferences();
+    if (CanHaveThreatList())
+         DeleteThreatList();
+    if (GetTypeId() == TYPEID_PLAYER)
+        getHostileRefManager().setOnlineOfflineState(false);
+    else
+        getHostileRefManager().deleteReferences();
     GetMotionMaster()->Clear(false);                    // remove different non-standard movement generators.
 }
